requirements.txt
# Основные зависимости
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Werkzeug==2.0.1
Flask-Migrate==4.0.7
Flask-WTF==0.15.1
WTForms==3.2.1

# Telegram Bot
python-telegram-bot==13.4
APScheduler==3.6.3

# Утилиты
python-dotenv==0.19.0
python-dateutil==2.9.0
pytz==2024.2
click==8.1.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2

# Безопасность и криптография
cryptography==3.4.8
oauthlib==2.1.0
requests==2.31.0
requests-oauthlib==1.1.0

# Развертывание
gunicorn==20.1.0

# База данных
psycopg2-binary==2.9.2
SQLAlchemy==1.4.23

# Дополнительные зависимости
numpy==2.2.0
pandas==2.2.3
plotly==5.14.1
openpyxl==3.1.2

# Тестирование
pytest==7.4.3
pytest-cov==4.1.0
